+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - Implement `GetChatId` for `Update`.
+ - The `dialogue::enter()` function as a shortcut for `dptree::entry().enter_dialogue()`.
 
 ## 0.8.0 - 2022-04-18